pyproject.toml
[project]
name = "beambo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eurocodepy>=0.1.17",
    "jinja2>=3.1.6",
    "matplotlib>=3.10",
    "numpy>=2.2",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.67.0",
    "PyNiteFEA>=0.0.102",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "pyright>=1.1",
    "pre-commit>=4.5.1",
]

[tool.pytest.ini_options]
testpaths = ["test"]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E741"]

[tool.pyright]
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"
